+++ lib/surveyor/models/survey_translation_methods.rb
@@ -11,10 +11,6 @@ module Surveyor
         # Associations
         belongs_to :survey, optional: true
 
-        if defined? ActiveModel::MassAssignmentSecurity
-          attr_accessible *PermittedParams.new.survey_translation_attributes
-        end
-
         # Validations
         validates_presence_of :locale, :translation
         validates_uniqueness_of :locale, scope: :survey_id