+++ lib/surveyor/models/dependency_condition_methods.rb
@@ -10,11 +10,15 @@ module Surveyor
 
       included do
         # Associations
-        belongs_to :answer, required: false
-        belongs_to :dependency, required: false
-        belongs_to :dependent_question, foreign_key: :question_id, class_name: :question, required: false
-        belongs_to :question, required: false
-        attr_accessible *PermittedParams.new.dependency_condition_attributes if defined? ActiveModel::MassAssignmentSecurity
+        belongs_to :answer, optional: true
+        belongs_to :dependency, optional: true
+
+        # what is the difference between :dependency_question and :question?
+        belongs_to :dependent_question,
+          foreign_key: :question_id,
+          class_name: :question,
+          optional: true
+        belongs_to :question, optional: true
 
         # Validations
         validates_presence_of :operator, :rule_key
@@ -31,17 +35,36 @@ module Surveyor
       # Instance methods
       def to_hash(response_set)
         # all responses to associated question
-        responses = question.blank? ? [] : response_set.responses.where('responses.answer_id in (?)', question.answer_ids)
+        responses =
+          if question.blank?
+            []
+          else
+            response_set.responses.where('responses.answer_id in (?)', question.answer_ids)
+          end
+
         if operator.match /^count(>|>=|<|<=|==|!=)\d+$/
           op, i = operator.scan(/^count(>|>=|<|<=|==|!=)(\d+)$/).flatten
           # logger.warn({rule_key.to_sym => responses.count.send(op, i.to_i)})
-          return { rule_key.to_sym => (op == '!=' ? !responses.count.send('==', i.to_i) : responses.count.send(op, i.to_i)) }
-        elsif (operator == '!=') && (responses.blank? || responses.none? { |r| r.answer.id == answer.id })
+
+          rule_key_value =
+            if op == '!='
+              !responses.count.send('==', i.to_i)
+            else
+              responses.count.send(op, i.to_i)
+            end
+
+          return { rule_key.to_sym => rule_key_value }
+        elsif (operator == '!=') &&
+            (responses.blank? || responses.none? { |r| r.answer.id == answer.id })
           # logger.warn( {rule_key.to_sym => true})
           return { rule_key.to_sym => true }
         elsif response = responses.to_a.detect { |r| r.answer.id == answer.id }
           klass = response.answer.response_class
-          klass = 'answer' if as(klass).nil? # it should compare answer ids when the dependency condition *_value is nil
+          if as(klass).nil?
+            # it should compare answer ids when the dependency condition *_value is nil
+            klass = 'answer'
+          end
+
           case operator
           when '==', '<', '>', '<=', '>='
             # logger.warn( {rule_key.to_sym => response.as(klass).send(self.operator, self.as(klass))})
@@ -58,9 +81,10 @@ module Surveyor
       protected
 
       def validates_operator
-        errors.add(:operator, 'Invalid operator') unless
-          Surveyor::Common::OPERATORS.include?(operator) ||
+        unless Surveyor::Common::OPERATORS.include?(operator) ||
             operator&.match(/^count(<|>|==|>=|<=|!=)(\d+)/)
+          errors.add(:operator, 'Invalid operator')
+        end
       end
     end
   end