+++ lib/surveyor/models/skip_logic_methods.rb
@@ -11,14 +11,17 @@ module Surveyor
 
       included do
         # Associations
-        belongs_to :survey_section, inverse_of: :skip_logics, required: false
-        belongs_to :target_survey_section, foreign_key: :target_survey_section_id, class_name: 'SurveySection', required: false
+        belongs_to :survey_section, inverse_of: :skip_logics, optional: true
+        belongs_to :target_survey_section,
+          foreign_key: :target_survey_section_id,
+          class_name: 'SurveySection',
+          optional: true
         has_many :skip_logic_conditions, inverse_of: :skip_logic, dependent: :destroy
-        attr_accessible *PermittedParams.new.skip_logic_attributes if defined? ActiveModel::MassAssignmentSecurity
 
         # Validations
         validates_presence_of :rule
-        validates_format_of :rule, with: /\A(?:and|or|\)|\(|[A-Z]|\s)+\Z/ # TODO properly formed parenthesis etc.
+        # TODO properly formed parenthesis etc.
+        validates_format_of :rule, with: /\A(?:and|or|\)|\(|[A-Z]|\s)+\Z/
         validates_presence_of :survey_section
       end
 
@@ -32,11 +35,21 @@ module Surveyor
         # logger.debug "rexp: #{rgx.inspect}"
         # logger.debug "keyp: #{ch.inspect}"
         # logger.debug "subd: #{self.rule.gsub(rgx){|m| ch[m.to_sym]}}"
-        rgx = Regexp.new(skip_logic_conditions.map { |slc| ['a', 'o'].include?(slc.rule_key) ? "\\b#{slc.rule_key}(?!nd|r)\\b" : "\\b#{slc.rule_key}\\b" }.join('|')) # exclude and, or
+
+        rgx_values = skip_logic_conditions.map do |slc|
+          if ['a', 'o'].include?(slc.rule_key)
+            "\\b#{slc.rule_key}(?!nd|r)\\b"
+          else
+            "\\b#{slc.rule_key}\\b"
+          end
+        end.join('|')
+
+        rgx = Regexp.new(rgx_values) # exclude and, or
         eval(rule.gsub(rgx) { |m| ch[m.to_sym] })
       end
 
-      # A hash of the conditions (keyed by rule_key) and their evaluation (boolean) in the context of response_set
+      # A hash of the conditions (keyed by rule_key) and their evaluation (boolean)
+      # in the context of response_set
       def conditions_hash(response_set)
         hash = {}
         skip_logic_conditions.each { |slc| hash.merge!(slc.to_hash(response_set)) }