+++ spec/factories.rb
@@ -9,15 +9,15 @@ FactoryBot.define do
   sequence(:unique_survey_access_code) { |_n| "simple survey #{UUIDTools::UUID.random_create}" }
 
   factory :survey do |s|
-    s.title           'Simple survey'
-    s.description     'A simple survey for testing'
+    s.title { 'Simple survey' }
+    s.description { 'A simple survey for testing' }
     s.access_code     { FactoryBot.generate :unique_survey_access_code }
-    s.survey_version  0
+    s.survey_version { 0 }
   end
 
   factory :survey_translation do |t|
-    t.locale 'es'
-    t.translation %(title: "Un idioma nunca es suficiente"
+    t.locale { 'es' }
+    t.translation { %(title: "Un idioma nunca es suficiente"
   survey_sections:
     one:
       title: "Uno"
@@ -28,7 +28,7 @@ FactoryBot.define do
       text: "¿Cómo se llama Usted?"
       answers:
         name:
-          help_text: "Mi nombre es...")
+          help_text: "Mi nombre es...") }
   end
 
   sequence(:survey_section_display_order) { |n| n }
@@ -47,19 +47,19 @@ FactoryBot.define do
   factory :question do |q|
     q.association             :survey_section # s.survey_section_id       {}
     # q.question_group_id       {}
-    q.text                    'What is your favorite color?'
-    q.short_text              'favorite_color'
-    q.help_text               'just write it in the box'
-    q.pick                    :none
-    q.reference_identifier    { |me| "q_#{me.object_id}" }
+    q.text { 'What is your favorite color?' }
+    q.short_text { 'favorite_color' }
+    q.help_text { 'just write it in the box' }
+    q.pick { :none }
+    q.reference_identifier { |me| "q_#{me.object_id}" }
     # q.data_export_identifier  {}
     # q.common_namespace        {}
     # q.common_identifier       {}
-    q.display_order           FactoryBot.generate(:question_display_order)
+    q.display_order { FactoryBot.generate(:question_display_order) }
     # q.display_type            {} # nil is default
-    q.is_mandatory            false
+    q.is_mandatory { false }
     # q.display_width           {}
-    q.correct_answer_id       nil
+    q.correct_answer_id { nil }
   end
 
   factory :question_group do |g|
@@ -78,18 +78,18 @@ FactoryBot.define do
 
   factory :answer do |a|
     a.association               :question # a.question_id               {}
-    a.text                      'My favorite color is clear'
-    a.short_text                'clear'
-    a.help_text                 'Clear is the absense of color'
+    a.text { 'My favorite color is clear' }
+    a.short_text { 'clear' }
+    a.help_text { 'Clear is the absense of color' }
     # a.weight
     # a.response_class            {}
     # a.reference_identifier      {}
     # a.data_export_identifier    {}
     # a.common_namespace          {}
     # a.common_identifier         {}
-    a.display_order             { FactoryBot.generate :answer_display_order }
+    a.display_order { FactoryBot.generate :answer_display_order }
     # a.is_exclusive              {}
-    a.display_type              'default'
+    a.display_type { 'default' }
     # a.display_length            {}
     # a.custom_class              {}
     # a.custom_renderer           {}